requirements.txt
Werkzeug==2.0.1
Flask==2.0.1
numpy==1.26.4
pandas>=1.5.3
yfinance==0.1.70
python-dotenv==0.19.0
openai==0.27.0
--only-binary :all: 